% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot PCA}
\usage{
plot_pca(input1, proteinGroups)
}
\arguments{
\item{input1}{experimental design input}

\item{proteinGroups}{protein groups input}
}
\description{
This function loads files and plot PCA as output
}
