% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_demo.R
\name{plot_pca_demo}
\alias{plot_pca_demo}
\title{plot PCA}
\usage{
plot_pca_demo(fooinput1, fooinput2, ...)
}
\arguments{
\item{input1}{experimental design input}

\item{input2}{protein groups input}
}
\description{
This function loads files and plot PCA as output
}
